remove protections of NodeContextManager class and require users to call the delegate node directly.
Unfortunately, this means we won't benefit anymore from the source-level putting
it in our face that we should be thinking about whether we need to transfer the
state. But the way this was done before, Truffle and PE don't see that the node
is executed as part of an AST in which it is adopted, even though it is a final
field in the context manager classes.

The context managers are a nice idea, but until we figure out how to make the
intent clear to PE, we have to do without.