force classloading when we import and we are emulating jython.
It seems to me this is the only way to reliably ensure that when a Java class is
imported that is not yet loaded that it will be listed in
java.lang.Package.getPackages() so we can reliably detect Java packages. This is
unfortunate, but importing is already slow in Jython mode, so I guess it won't
hurt us any more than it already does.