Don't execute finally clauses on Java exceptions
Such exceptions signify interpreter crash and shouldn't be handed over
to the program, as it might suppress them. There's a internal setting
`CatchAllExceptions` which allows the Python code to catch Java
exceptions. When this setting is on, Java exceptions get wrapped in
Python exceptions and re-raised, thus the finally handlers are executed
as well.