Make iterating over hashes more convenient
This benefits dict.__repr__ and dictviews.

We now have a generic forEach message to implement, and injectInto is a
convenience method in terms of it. The HashingStorageLibrary now also returns a
final subclass of Iterable, and its iterator method a final subclass of
Iterator. This final subclass has boundaries on the next and hasNext calls, so
we avoid the interface calls during PE, and the implementations of
HashingStorageLibrary don't have to provide those boundaries. We also now get
the iterator in the node call site when creating a dict view.