Pass shapes in all object constructors to make them mostly constant for PE
make class shapes constant more often and report polymorphism if they aren't

Create the DynamicObject already in the nodes

Instead of passing the shape to PythonObject constructors, we now pass a freshly
created DynamicObject. The reason is that Shape.newInstance may legitimately
become megamorphic and thus we cannot inline and want that behind a
TruffleBoundary. This is mostly done in the PythonObjectFactory node anyway, so
its not a big hassle to use.