Tweak performance of async handler
* Let async handler trigger on function return and only every Short.MAX_VALUE loop iterations

* This also removes the transferToInterpreter in the async action trigger, and
instead uses a TruffleBoundary. Additionally, we now let async handlers that
need the frame always walk the stack. These handlers aren't common, and this is
slightly better than transferring to the interpreter all the time. We also add a
branch profile for the async handler and an option to disable it, to gauge its
impact.

* The compiler generates much better code when the async actions are in the beginning

* Further tweak the async handler triggers

* Transfer to interpreter for async actions anyway. it creates faster code