Use SecureRandom for os.urandom
The Random PRNG does not produce random numbers sufficient for
cryptographic use, as os.urandom specifies it should. Instead we replace
it with an instance of SecureRandom.

Because creating SecureRandom instances consumes a significant amount of
resources, we store one instance in a static field to share amongst all
URandomNodes. We initialize it only on the first call to urandom to
avoid incurring the cost on loading the class.