Put BuiltinMethodInfo objects also into slots of PythonManagedClass
BuiltinMethodInfo objects are context & language independent and can be
freely cached in the AST. Moreover, they provide direct way to call the
underlying builtin node without the normal calling machinery (as they
contain reference to its node factory).

BuiltinMethodInfo instances are cached and unique per (builtin node
factory, declaring type), which alows to use identity comparison in
guards, whithout indirection.

Specializations in CallBinaryMethodNode and co. that handle BuiltinMethodInfo
either insert the builtin node into the AST, or if the cache overflows
the generic specialization uses IndirectCallContext and TruffleBounary
call with null frame.

With this change LookupAttributeInMRONode is reverted to the state on
master: it does never lookup in slots and it is still always guaranteed
to return a runtime object.

When the caller can handle BuiltinMethodInfo, it should use
LookupCallableSlotInMRONode instead. LookupAndCallBinaryNode and co. are
updated to use LookupCallableSlotInMRONode where possible.