Force inner contexts to have their own language instance
This ensures that the singleContext assumption gets invalidated either
at the beginning before any AST is executed or never. This commit also
adds assertions that check this assumption.

Followup work will replace the singleContextAssumption with just the
singleContext field, which is enough and can be checked only once, e.g.,
only when activating a specialization on the slow path.