Add utf-32 character support for csv parsing and writing.
Python supports utf-32 characters naturally. As Java characters are limited to utf-16 this became a problem when
special characters which have to be compared to during parsing and writing (e.g. csv delimiters) were
set to utf-32 characters. E.g. it was previously not possible to use a smiley  (🙂) as a delimite in a csv.

To implement support for utf-32 characters parsing and writing of csvs is now done using code points instead
of characters.

Additionally tests for utf-32 characters were added.