Refactor: Use `.equals()` instead of equality operator (==) for comparing string constant marker.
As `NOT_SET` is a string constant marker that we set to mark not set values, e.g. for an escape char, it would
be possible to compare for it via the equality operator. However, the python style check doesn't accept that
and I could not find a specific `@SuppressWarnings` annotation for this use case, therefore I decided to switch
the checks to `.equals()`.