fix IsSubtypeNode for variable subtypes and constant basetype
The isSubMro method tried to be clever and only check if the basetype is at a
single position in the subtype's MRO. This works fine if all classes from the
subtype up to (but not including) the basetype class only have a single
base. But if anything in between has multiple bases, the MRO will have extra
elements in between.

Example: _io.UnsupportedOperation inherits from ValueError and OSError, leading
to this MRO:
   <class 'io.UnsupportedOperation'>
   <class 'OSError'>
   <class 'ValueError'>
   <class 'Exception'>
   <class 'BaseException'>
   <class 'object'>

OSError has this:
   <class 'OSError'>
   <class 'Exception'>
   <class 'BaseException'>
   <class 'object'>

The diff is two, previously the IsSubtypeNode might now check at index 2 of the
subtype's MRO, and find ValueError to be not the same as OSError. Thus the need
to loop from index 0 to index 2.