Fix: REPL completion adds trailing backslash after keyword completion.
The REPL completer implemented by `rlcompleter.py` adds a trailing whitespace to keywords (except `finally` and `try`).
This makes sense as nothing else can follow those keywords, however, the current GraalPython completer implementation
via `JLineConsoleHandler` uses a DefaultParser for parsing completion candidates which escapes whitespace characters.
Therefore a trailing whitespace got replaced by '\ ' leading to a wrong completion. As a fix we overwrite the
`isDelimiterChar` method from the DefaultParser to never count the last character of a character sequence as a delimiter.
This works because we otherwise only use this check to split a line into it's words and in this case the last
character of the line also cannot be a delimiter.