Ensure co_consts doesn't provide python access to a raw java array
At some point in jinja2.loaders, the code object for the function load gets hashed. Since
this function contains the list view ['root'], it used to be placed into co_consts as
a singleton Object[] containing a TruffleString. This then made the hash crash. This
way it is stored as a tuple and can get hashed normally, fixing the issue