[GR-23351][GR-28624][GR-42663] remove intrinsified zipimport module
Since we cannot use java.zip.ZipFile (as that would escape the truffle
sandbox), we would have to replicate the ZipFile code to deal with all zip
features to a) pass the tests and b) improve performance for random access to
files in the zipimport case. Switching to the stdlib zipimport.py (as CPython
has done) will ensure a) and for larger zip files even b) immediately, with the
option of improving peak performance later.