Add rudimentary support for _tkinter in GraalPy
This commit adds the _tkinter module to the list of built-in modules to allow use of tkinter.
Before use, the bindings need to be built using tklib_build.py.
The code was modified from existing code in PyPy's main branch.

Right now, opening a window under Linux works, but still crashes on interacting with UI elements such as buttons.
It seems like there is a bug in the `cffi` module causing a SegFault.
macOS complains about other threads accessing the main window and the build script has some trouble finding the correct library for Tcl and Tk, as macOS comes with an older version preinstalled.
Windows was not tested.

Co-authored-by: Margarete Dippel <margarete.dippel@student.hpi.de>
Co-authored-by: Tim Felgentreff <tim.felgentreff@oracle.com>