Introduce registration of custom python classes associated to Java Classes
Currently, this enables the registration for custom python classes that should
be associated with certain Java classes.

The registry is part of a PythonContext and is currently indexed by String keys,
which are the java class names for which python classes have been registered.
Registration takes place in a respective PolyglotModuleBuiltin and the lookup
happens in the GetClassNode#getForeign implementation.

Usage in python: `register_java_interop_type("{java_class_name}", pythonClass)`

In the future, also an annotation-style registration way should be implemented.