Use CallBinaryMethodNode for dispatching __getattribute__ callable
It is a common pattern that Python class defines only __getattr__, in which
case the TpSlotGetAttrPython will wrap the inherited builtin __getattribute__
and the user defined __getattr__ callable.

Normally, slots wrapping Python callables do not wrap builtins. This is
impossible for slots wrapping one callable, because we unwrap them in
fixupSlotDispatchers (like CPython). For other slots wrapping 2 callables it is
probably not as common as for tp_getattro. Therefore the dispatching code for
those nodes, does not have a fast-path for builtin callables. This PR adds such
fast-path to CallManagedSlotGetAttrNode.