Clean up code around registering replacement wrappers
All replacements must be freed somehow, and they can only be free'd when the
Java side says it's ok to do so, either when the managed object dies or the
context shuts down. So any replacement wrapper that is not simply free'd
directly (like handles) needs to be free'd by calling its type's tp_dealloc.